Project Porg/backend/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==5.0.1
anyio==3.6.2
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
azure-cognitiveservices-speech==1.25.0
beautifulsoup4==4.11.1
blinker==1.6.2
Brotli==1.0.9
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
contourpy==1.0.7
cryptography==39.0.0
cssselect==1.2.0
cycler==0.11.0
dataclasses-json==0.5.7
decorator==4.0.11
dnspython==2.3.0
elevenlabslib==0.4.0
fake-useragent==1.1.3
fastapi==0.89.1
ffmpeg==1.4
ffmpy==0.3.0
filelock==3.12.0
filetype==1.0.5
Flask==2.2.2
Flask-Cors==4.0.0
fonttools==4.39.4
frozendict==2.3.4
frozenlist==1.3.3
fsspec==2023.5.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-images-download @ git+https://github.com/Joeclinton1/google-images-download.git@e91e6a38ad654877f59edc7894822f5319d75df1
Google-Images-Search==1.4.6
googleapis-common-protos==1.59.0
gradio==3.32.0
gradio_client==0.2.5
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.14.1
hyperframe==6.0.1
icoextract==0.1.4
icrawler==0.6.6
idna==3.4
imageio==2.28.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
insta-scrape==2.1.2
instagrapi==1.17.6
install==1.3.5
instareels==0.1.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kaitaistruct==0.10
keyboard==0.13.5
kiwisolver==1.4.4
langchain==0.0.165
Levenshtein==0.20.9
linkify-it-py==2.0.2
loguru==0.7.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
MouseInfo==0.1.3
moviepy==1.0.3
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.1
oauth2client==4.1.3
oauthlib==2.0.1
openai==0.27.2
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.68
orjson==3.8.14
outcome==1.2.0
packaging==23.1
pandas==1.5.3
parse==1.19.0
pefile==2023.2.7
Pillow==9.5.0
pinecone-client==2.2.1
playsound==1.3.0
proglog==0.1.10
progressbar==2.5
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.7
pydivert==2.1.0
pydub==0.25.1
pyee==8.2.2
pyfiglet==0.8.post1
pygame==2.3.0
PyGetWindow==0.0.9
pygi==1.0.3
Pygments==2.15.1
PyMsgBox==1.0.9
PyOpenGL==3.1.6
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
pyppeteer==1.0.2
pyquery==2.0.0
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.20.9
python-magic-bin==0.4.14
python-multipart==0.0.6
python-resize-image==1.1.20
pyttsx3==2.90
pytube==15.0.0
pytweening==1.0.4
pytz==2022.7.1
pywin32==305
PyYAML==6.0
rapidfuzz==2.13.7
regex==2023.5.5
requests==2.29.0
requests-html==0.10.0
requests-oauthlib==0.7.0
rsa==4.9
scipy==1.10.1
selenium==4.10.0
selenium-wire==5.1.0
semantic-version==2.10.0
simple-image-download==0.5
six==1.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
SQLAlchemy==2.0.10
starlette==0.22.0
tenacity==8.2.2
termcolor==1.1.0
tiktoken==0.4.0
toolz==0.12.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
typing==3.7.4.3
typing-inspect==0.8.0
typing_extensions==4.4.0
uc-micro-py==1.0.2
undetected-chromedriver==3.5.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
w3lib==2.1.1
webdriver-manager==3.8.6
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
win32-setctime==1.1.0
windows-curses==2.3.1
wsproto==1.2.0
yarl==1.8.2
yfinance==0.2.9
zipp==3.15.0
zstandard==0.21.0
